Move common functionality to utilities.py
